Complete PSTA with analysis, throws outofmem error for full dataset

+++ PeriodicTopicBehaviors/src/main/java/edu/ntnu/app/psta/Variable.java
@@ -2,6 +2,8 @@ package edu.ntnu.app.psta;
 
 public interface Variable {
     boolean update();
+
     void setVars(VariableList p1, VariableList p2);
-    double get(int... values);
+
+    Float get(int... values);
 }
